requirements.txt
aiohttp
pyotp
python-dotenv
cachetools
asyncio
motor
pymongo
pwinput
aiofiles